с функ-комментами
