+++ main.js
@@ -1,14 +1,11 @@
-import { renderComments } from "./render.js";
 import { fetchAndCommentsRender } from "./fetchnrender.js";
-import { postComment } from "./postcomment.js";
+import { handlePostClick } from "./handlepostclick.js";
 
 // Объявление переменных
 export const nameInputElement = document.getElementById("name-input");
 export const textAreaElement = document.getElementById("text-input");
 export const addFormElement = document.querySelector(".add-form");
 export const addCommentElement = document.getElementById("add-comment");
-
-const buttonElement = document.getElementById("write-button");
 
 // Структура данных для комментариев
 // let commentsData = [];
@@ -42,58 +39,60 @@ const buttonElement = document.getElementById("write-button");
 //         });
 // };
 
-// функция проверки полей и обработка кодов API
-const handlePostClick = (commentsData) => {
+// // функция проверки полей и обработка кодов API
+// const handlePostClick = (commentsData) => {
 
-    nameInputElement.classList.remove("error");
-    textAreaElement.classList.remove("error");
+//     nameInputElement.classList.remove("error");
+//     textAreaElement.classList.remove("error");
 
-    if (!nameInputElement.value || nameInputElement.value.trim().length === 0) {
-        nameInputElement.classList.add("error");
-        return;
-    }
+//     if (!nameInputElement.value || nameInputElement.value.trim().length === 0) {
+//         nameInputElement.classList.add("error");
+//         return;
+//     }
 
-    else if (!textAreaElement.value || textAreaElement.value.trim().length === 0) {
-        textAreaElement.classList.add("error");
-        return;
-    }
+//     else if (!textAreaElement.value || textAreaElement.value.trim().length === 0) {
+//         textAreaElement.classList.add("error");
+//         return;
+//     }
 
-    // Скрыть форму добавления и показать сообщение о добавлении
-    addFormElement.style.display = "none";
-    addCommentElement.textContent = "Комментарий добавляется...";
-    addCommentElement.style.display = "block";
+//     // Скрыть форму добавления и показать сообщение о добавлении
+//     addFormElement.style.display = "none";
+//     addCommentElement.textContent = "Комментарий добавляется...";
+//     addCommentElement.style.display = "block";
 
-    postComment(textAreaElement.value, nameInputElement.value)
+//     postComment(textAreaElement.value, nameInputElement.value)
 
-        .catch((error) => {
+//         .catch((error) => {
 
-            // Обработка ошибки сервера
-            if (error.message === "Сервер упал") {
-                handlePostClick();
-            }
+//             // Обработка ошибки сервера
+//             if (error.message === "Сервер упал") {
+//                 handlePostClick();
+//             }
 
-            // Обработка ошибки пользователя
-            else if (error.message === "Ошибочный запрос") {
-                alert("Имя и комментарий должны быть не короче 3 символов");
-                showAddForm();
-                return;
-            }
+//             // Обработка ошибки пользователя
+//             else if (error.message === "Ошибочный запрос") {
+//                 alert("Имя и комментарий должны быть не короче 3 символов");
+//                 showAddForm();
+//                 return;
+//             }
 
-            else {
-                alert("Кажется, у вас сломался интернет, попробуйте позже");
-                showAddForm();
-            }
-        });
+//             else {
+//                 alert("Кажется, у вас сломался интернет, попробуйте позже");
+//                 showAddForm();
+//             }
+//         });
 
-    renderComments(commentsData);
-};
+//     renderComments(commentsData);
+// };
+
+// buttonElement.addEventListener('click', handlePostClick);
+
+// // Показ заполненной формы
+// function showAddForm() {
+//     addFormElement.style.display = "flex";
+//     addCommentElement.style.display = "none";
+// }
+const buttonElement = document.getElementById("write-button");
 
 buttonElement.addEventListener('click', handlePostClick);
-
-// Показ заполненной формы
-function showAddForm() {
-    addFormElement.style.display = "flex";
-    addCommentElement.style.display = "none";
-}
-
 fetchAndCommentsRender();
